refactor(plugin-registry): extract dev-plugin settings wiring to a container
The settings surface had a substantial inline component pulling
`usePluginManager`, `useSettingsState`, `useAtomValue(devPluginIds)`,
and the add/enable/remove callbacks. Move it to
`containers/RegistrySettingsContainer` to match the convention used by
the other registry articles (PluginsArticle, PluginRegistryArticle,
etc.) and keep the Surface registration thin.

No behavior change.
