+++ packages/plugins/plugin-registry/src/containers/RegistrySettingsContainer/index.ts
@@ -1,0 +1,5 @@
+//
+// Copyright 2026 DXOS.org
+//
+
+export { RegistrySettingsContainer as default } from './RegistrySettingsContainer';
